requirements.txt
certifi==2020.6.20
cffi==1.14.0
cryptography==2.9.2
future==0.18.2
pycparser==2.20
python-telegram-bot==11.1.0
six==1.15.0
